backend/requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0
jsonschema>=4.0.0

# LangGraph and LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0

# GAM-style memory: BM25 + Vector search
rank-bm25>=0.2.2
sentence-transformers>=2.2.0
numpy>=1.24.0
faiss-cpu>=1.7.4

# Training and RL
torch>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# CLI
click>=8.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# OpenAI (for real model training)
openai>=1.0.0

# Agent Lightning (for RL training + official APO dependencies)
agentlightning[apo]>=0.1.0

# Isolated API execution sandbox
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0

# GAM (General Agentic Memory) dependencies
sentence-transformers>=2.2.0
rank-bm25>=0.2.2

# Optional OSS integration stack (v2 architecture)
openapi-core>=0.19.0
openapi-spec-validator>=0.7.0
hypothesis>=6.0.0
# schemathesis, pact, locust, and testcontainers are optional at runtime.
# The agent detects availability and reports ready/skipped status in `oss_tooling`.
